HEAD:(zeroc-ice/ice@776437e2ce) Reduce usage of property regex matching (zeroc-ice/ice#2122)
Updates the generated property code to include a boolean indicating if the property uses regular expressions for matching. It also removes the case sensitivity check on the property names in favor of a check on the prefix.

Now, if a property does not require regex we just do a plain string comparison.

Closes zeroc-ice/ice#2108 and zeroc-ice/ice#2106