HEAD:(zeroc-ice/ice@4bdf7c33d2) Add language specification to property names (zeroc-ice/ice#2795)
This is the first step in implementing zeroc-ice/ice#2746. Namely,

Properties in PropertyNames.xml now require an additional attribute languages="...". This specifies the languages which support this property. Languages which do not support the property do not generate an entry for it. As a consequence they'll warnings will be issued they you try to use an unsupported property.

languages is a comma separated list of cpp, csharp, java, js, or a all (by itself).

Reworked makeprops to delay generation until after we've parsed all the properties. We need to know ahead of time if an entire section will be empty or not.