HEAD:(zeroc-ice/ice@55308b5103) Better C++ property validation (zeroc-ice/ice#2838)
Part of zeroc-ice/ice#2746, starting with C++. Reworked the generated PropertyNames code to create property arrays for the property classes and updated that properties that reference a class to refer to these.

There are several benefits: less generated code, property classes can now be self referential, and we can use the new property class arrays to validate user property classes (done for OAs so far) and to get defaults (not implemented yet).