HEAD:(zeroc-ice/ice@64991677b3) Make (some) Ice properties opt-in (C++ only) (zeroc-ice/ice#2983)
Updates the generated PropertyNames to indicate if a property array is opt-in. Properties whose prefix is opt-in can now no longer be set unless the corresponding Property object is constructed with an opt-in list containing its prefix.

I've also renamed UnknownPropertyException to PropertyException so we can use it for a few more things.

Part of zeroc-ice/ice#2855